src/main/java/com/alexfehd/customerservicechat/controller/ChatController.java
package com.alexfehd.customerservicechat.controller;

import com.alexfehd.customerservicechat.entity.ChatMessage;
import com.alexfehd.customerservicechat.entity.ChatNotification;
import com.alexfehd.customerservicechat.entity.User;
import com.alexfehd.customerservicechat.service.ChatMessageService;
import com.alexfehd.customerservicechat.service.ChatRoomService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.messaging.handler.annotation.MessageMapping;
import org.springframework.messaging.handler.annotation.Payload;
import org.springframework.messaging.simp.SimpMessagingTemplate;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;

import java.util.List;

@Controller
@RequiredArgsConstructor
public class ChatController {

    private final SimpMessagingTemplate messagingTemplate;
    private final ChatMessageService chatMessageService;
    private final ChatRoomService chatRoomService;

    @MessageMapping("/chat")
    public void processMessage(@Payload ChatMessage chatMessage) {
        ChatMessage savedMsg = chatMessageService.save(chatMessage);
        messagingTemplate.convertAndSendToUser(
                chatMessage.getRecipientId(), "/queue/messages",
                new ChatNotification(
                        savedMsg.getId(),
                        savedMsg.getSenderId(),
                        savedMsg.getRecipientId(),
                        savedMsg.getContent()
                )
        );
    }

    @GetMapping("/messages/{senderId}/{recipientId}")
    public ResponseEntity<List<ChatMessage>> findChatMessages(@PathVariable String senderId,
                                                 @PathVariable String recipientId) {
        return ResponseEntity
                .ok(chatMessageService.findChatMessages(senderId, recipientId));
    }

    @GetMapping("/messages/{recipientId}")
    public ResponseEntity<List<User>> findAllOpenChats( @PathVariable String recipientId) {

        return ResponseEntity
                .ok(chatRoomService.findAllUsersWithOpenChatRooms(recipientId));
    }


}
